This commit adds two parameters: `--data-transfer-limit` and `--delay-transfer-limit-check`. The former sets a limit on the amount of data to be transferred (does not account for delta transfer), and the latter allows the limit to be exceeded in cases where the limit has not yet been reached, but the next file will exceed it. For example, if the limit is 100MiB and 99MiB have transferred, and the next file is 10MiB, the delay parameter will allow that file to be transferred. In this example, without the delay parameter, rsync will stop at 99MiB.

+++ generator.c
@@ -99,6 +99,23 @@ extern char *tmpdir;
 extern char *basis_dir[MAX_BASIS_DIRS+1];
 extern struct file_list *cur_flist, *first_flist, *dir_flist;
 extern filter_rule_list filter_list, daemon_filter_list;
+static int64 stop_after_data_transferred_data = 0;
+extern int data_transfer_limit_reached;
+extern int64 data_transfer_limit;
+extern int delay_transfer_limit_check;
+static int data_transfer_limit_msg_sent = 0;
+
+static int data_transfer_limit_would_be_exceeded(struct file_struct *file)
+{
+	if (data_transfer_limit < 0)
+		return 0;
+
+	if (delay_transfer_limit_check)
+		return stop_after_data_transferred_data >= data_transfer_limit;
+
+	return stop_after_data_transferred_data + F_LENGTH(file) > data_transfer_limit;
+}
+
 
 int maybe_ATTRS_REPORT = 0;
 int maybe_ATTRS_ACCURATE_TIME = 0;
@@ -1240,6 +1257,15 @@ static void recv_generator(char *fname, struct file_struct *file, int ndx,
 	if (DEBUG_GTE(GENR, 1))
 		rprintf(FINFO, "recv_generator(%s,%d)\n", fname, ndx);
 
+
+	if (data_transfer_limit_reached) {
+#ifdef SUPPORT_HARD_LINKS
+		if (preserve_hard_links && F_IS_HLINKED(file))
+			handle_skipped_hlink(file, itemizing, code, f_out);
+#endif
+		return;
+	}
+
 	if (list_only) {
 		if (is_dir < 0
 		 || (is_dir && !implied_dirs && file->flags & FLAG_IMPLIED_DIR))
@@ -1920,6 +1946,18 @@ static void recv_generator(char *fname, struct file_struct *file, int ndx,
 		rprintf(FINFO, "generating and sending sums for %d\n", ndx);
 
   notify_others:
+	if (data_transfer_limit_would_be_exceeded(file)) {
+		rprintf(FINFO,
+			"recv_generator: transferring %s would exceed data transfer limit. Not transferred; stopping all remaining transfers.\n",
+			fname);
+		data_transfer_limit_reached = 1;
+		if (!data_transfer_limit_msg_sent) {
+			send_msg_int(MSG_DATA_LIMIT_REACHED, 0);
+			io_flush(MSG_FLUSH);
+			data_transfer_limit_msg_sent = 1;
+		}
+		goto cleanup;
+	}
 	if (remove_source_files && !delay_updates && !phase && !dry_run)
 		increment_active_files(ndx, itemizing, code);
 	if (inc_recurse && (!dry_run || write_batch < 0))
@@ -1941,6 +1979,8 @@ static void recv_generator(char *fname, struct file_struct *file, int ndx,
 			fuzzy_file ? fuzzy_file->basename : NULL);
 		free_stat_x(&real_sx);
 	}
+
+	stop_after_data_transferred_data += F_LENGTH(file);
 
 	if (!do_xfers) {
 #ifdef SUPPORT_HARD_LINKS
@@ -2164,8 +2204,7 @@ void check_for_finished_files(int itemizing, enum logcode code, int check_redo)
 			continue;
 		}
 #endif
-
-		if (check_redo && (ndx = get_redo_num()) != -1) {
+		if (!data_transfer_limit_reached && check_redo && (ndx = get_redo_num()) != -1) {
 			OFF_T save_max_size = max_size;
 			OFF_T save_min_size = min_size;
 			csum_length = SUM_LENGTH;
@@ -2360,7 +2399,7 @@ void generate_files(int f_out, const char *local_name)
 		}
 	} while ((cur_flist = cur_flist->next) != NULL);
 
-	if (delete_during)
+	if (delete_during && !data_transfer_limit_reached)
 		delete_in_dir(NULL, NULL, dev_zero);
 	phase++;
 	if (DEBUG_GTE(GENR, 1))
@@ -2415,9 +2454,9 @@ void generate_files(int f_out, const char *local_name)
 	info_levels[INFO_FLIST] = save_info_flist;
 	info_levels[INFO_PROGRESS] = save_info_progress;
 
-	if (delete_during == 2)
+	if (delete_during == 2 && !data_transfer_limit_reached)
 		do_delayed_deletions(fbuf);
-	if (delete_after && !solo_file && file_total > 0)
+	if (delete_after && !solo_file && file_total > 0 && !data_transfer_limit_reached)
 		do_delete_pass();
 
 	if (max_delete >= 0 && skipped_deletes) {
